git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@11874 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ linalg/src/LinearSystem.cpp
@@ -210,7 +210,7 @@ LinearSystem::LinearSystem(Vec residualVector, Mat jacobianMatrix)
          *  Assuming that mLhsMatrix was created with PetscTools::SetupMat, the value
          *  below should be equivalent to what was used as preallocation in that call.
          */
-        mRowPreallocation = matrix_info.nz_allocated / mSize;
+        mRowPreallocation = (unsigned) matrix_info.nz_allocated / mSize;
     }
     assert(!mRhsVector || !mLhsMatrix || vec_size == mat_size);
